converted double to int

+++ src/wrappersc.c
@@ -31,7 +31,7 @@ void F77_NAME(pmodel_f)(
     double *latitude,
     double *altitude,
     double *whc,
-    double *temp_home,
+    int    *temp_home,
     int    *nt,
     double *par,
     double *forcing,
